Azur ai/requirements.txt
streamlit
azure-cognitiveservices-vision-computervision
azure-cognitiveservices-vision-face
azure-storage-blob
Pillow
requests
opencv-python-headless
numpy
